Renamed AppointmentBlockVaos to AppointmentBlock and deleted old AppointmentBlock.
