+++ src/applications/check-in/components/AppointmentBlock.jsx
@@ -1,141 +1,68 @@
-import React from 'react';
+import React, { useMemo } from 'react';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
-import { locationShouldBeDisplayed } from '../utils/appointment';
+// eslint-disable-next-line import/no-unresolved
+import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
+
+import { createAnalyticsSlug } from '../utils/analytics';
+import AppointmentListItemVaos from './AppointmentDisplay/AppointmentListItemVaos';
+import { makeSelectApp } from '../selectors';
+import { useFormRouting } from '../hooks/useFormRouting';
+import { APP_NAMES } from '../utils/appConstants';
+import { getAppointmentId, sortAppointmentsByStartTime } from '../utils/appointment';
 
 const AppointmentBlock = props => {
-  const { appointments, page } = props;
+  const { appointments, page, router } = props;
+  const selectApp = useMemo(makeSelectApp, []);
+  const { app } = useSelector(selectApp);
   const { t } = useTranslation();
-
   const appointmentsDateTime = new Date(appointments[0].startTime);
 
-  const infoBlockMessage = appointment => {
-    if (appointment?.kind === 'phone') {
-      if (page === 'confirmation') {
-        return t('your-provider-will-call-you-at-your-appointment-time');
-      }
-      return t('your-provider-will-call-you');
-    }
-    return t('please-bring-your-insurance-cards-with-you-to-your-appointment');
+  const { jumpToPage } = useFormRouting(router);
+
+  const handleDetailClick = (e, appointment) => {
+    e.preventDefault();
+    recordEvent({
+      event: createAnalyticsSlug('details-link-clicked', 'nav', app),
+    });
+    jumpToPage(`appointment-details/${getAppointmentId(appointment)}`);
   };
+
+  const sortedAppointments = sortAppointmentsByStartTime(appointments);
 
   return (
     <div>
-      <p
-        className="vads-u-font-family--serif"
-        data-testid="appointment-day-location"
-      >
-        {t('your-appointments-on-day', {
-          count: appointments.length,
-          day: appointmentsDateTime,
-        })}
-      </p>
+      {app === APP_NAMES.PRE_CHECK_IN ? (
+        <p
+          className="vads-u-font-family--serif"
+          data-testid="appointment-day-location"
+        >
+          {t('your-appointments-on-day', {
+            count: appointments.length,
+            day: appointmentsDateTime,
+          })}
+        </p>
+      ) : (
+        <p className="vads-u-font-family--serif" data-testid="date-text">
+          {t('here-are-your-appointments-for-today', { date: new Date() })}
+        </p>
+      )}
+
       <ol
-        className="vads-u-border-top--1px vads-u-margin-bottom--4 check-in--appointment-list"
+        className="vads-u-border-top--1px vads-u-margin-bottom--4 check-in--appointment-list appointment-list"
         data-testid="appointment-list"
       >
-        {appointments.map((appointment, index) => {
-          const appointmentDateTime = new Date(appointment.startTime);
-          const clinic = appointment.clinicFriendlyName
-            ? appointment.clinicFriendlyName
-            : appointment.clinicName;
+        {sortedAppointments.map(appointment => {
           return (
-            <li
-              key={index}
-              className="vads-u-border-bottom--1px check-in--appointment-item"
-              data-testid="appointment-list-item"
-            >
-              <div className="check-in--appointment-summary vads-u-margin-bottom--2 vads-u-margin-top--2">
-                <div className="vads-u-margin-right--1 check-in--label">
-                  <i
-                    aria-label="Appointment type"
-                    className={`fas ${
-                      appointment?.kind === 'phone' ? 'fa-phone' : 'fa-building'
-                    }`}
-                    aria-hidden="true"
-                  />
-                </div>
-                <div
-                  data-testid="appointment-type-label"
-                  className="check-in--value vads-u-font-weight--bold"
-                >
-                  {appointment?.kind === 'phone'
-                    ? t('phone-call')
-                    : t('in-person')}
-                </div>
-                {appointment?.kind !== 'phone' && (
-                  <>
-                    <div className="check-in--label vads-u-margin-right--1">
-                      Facility:
-                    </div>
-                    <div
-                      className="check-in--value"
-                      data-testid="facility-name"
-                    >
-                      {appointment.facility}
-                    </div>
-                  </>
-                )}
-                <div className="check-in--label vads-u-margin-right--1">
-                  {t('time')}:
-                </div>
-                <div className="check-in--value" data-testid="appointment-time">
-                  {t('date-time', { date: appointmentDateTime })}
-                </div>
-                <div className="check-in--label vads-u-margin-right--1">
-                  {t('type-of-care')}:
-                </div>
-                <div className="check-in--value" data-testid="type-of-care">
-                  {appointment.clinicStopCodeName
-                    ? appointment.clinicStopCodeName
-                    : t('VA-appointment')}
-                </div>
-                {appointment.doctorName && (
-                  <>
-                    <div className="check-in--label vads-u-margin-right--1">
-                      {t('provider')}:
-                    </div>
-                    <div className="check-in--value" data-testid="provider">
-                      {appointment.doctorName}
-                    </div>
-                  </>
-                )}
-                <div className="check-in--label vads-u-margin-right--1">
-                  {t('clinic')}:
-                </div>
-                <div
-                  className="check-in--value"
-                  data-testid="appointment-clinic"
-                >
-                  {clinic}
-                </div>
-                {locationShouldBeDisplayed(appointment) && (
-                  <>
-                    <div className="check-in--label vads-u-margin-right--1">
-                      {t('location')}:
-                    </div>
-                    <div
-                      className="check-in--value"
-                      data-testid="clinic-location"
-                    >
-                      {appointment.clinicLocation}
-                    </div>
-                  </>
-                )}
-              </div>
-              {page === 'confirmation' || appointment?.kind === 'phone' ? (
-                <va-alert
-                  background-only
-                  show-icon
-                  data-testid="appointment-message"
-                  class="vads-u-margin-bottom--2"
-                >
-                  <div>{infoBlockMessage(appointment)}</div>
-                </va-alert>
-              ) : (
-                ''
-              )}
-            </li>
+            <AppointmentListItemVaos
+              key={`${appointment.appointmentIen}-${appointment.stationNo}`}
+              appointment={appointment}
+              page={page}
+              goToDetails={handleDetailClick}
+              app={app}
+              router={router}
+            />
           );
         })}
       </ol>
@@ -146,6 +73,7 @@ const AppointmentBlock = props => {
 AppointmentBlock.propTypes = {
   appointments: PropTypes.array.isRequired,
   page: PropTypes.string.isRequired,
+  router: PropTypes.object,
 };
 
 export default AppointmentBlock;
